requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
browserforge==1.2.3
camoufox==0.4.11
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cssselect==1.3.0
curl_cffi==0.13.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
language-tags==1.2.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgspec==0.19.0
numpy==2.3.3
orjson==3.11.3
outcome==1.3.0.post0
patchright==1.55.2
platformdirs==4.4.0
playwright==1.55.0
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pymongo==4.15.1
PySocks==1.7.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
scrapling==0.3.5
screeninfo==0.8.1
selenium==4.35.0
sentry-sdk==2.39.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
tldextract==5.3.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.14.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
